Tweak `LiveContext` API
Discussion with @matbryan52: use builder pattern for proper
autocompletion support in IDE and hopefully jupyter notebooks. It also
allows to properly separate parameters that end up being passed to the
`Acquisition` object, and those used to construct the builder.

This adds `LiveContext.make_connection` and
`LiveContext.make_acquisition` methods, which allow to create
differet acquisitions and connection instances in a type-safe way.
It removes the previous `LiveContext.prepare_from_pending` method,
as `LiveContext.make_acquisition` works for that, too.
